wip: get console running
