+++ apps/console/bindings.d.ts
@@ -5,4 +5,5 @@ declare global {
   const SESSION_SECRET: string
   const COOKIE_DOMAIN: string
   const PASSPORT_URL: string
+  const STORAGE_NAMESPACE: string
 }
