+++ platform/address/src/core.ts
@@ -43,6 +43,9 @@ export default class Core extends DurableObject<
     const account = await this.storage.get<AccountURN | undefined>('account')
     if (account && !AccountURNSpace.is(account)) {
       const urn = AccountURNSpace.urn(account)
+      if (!AccountURNSpace.is(urn)) {
+        throw 'address.getAccount: invalid account'
+      }
       await this.setAccount(urn)
       return urn
     } else {
